[5.2] Admin Module Latest (#43216)
* [5.2] Admin Module Latest

The admin module Articles - Latest displays a list of recently added content. This is really useful on the dashboards. (especially when used together with the author filter)

The list can be sorted by created or modified date but it will only ever display the created date.

This PR changes the date column to display created or modified depending on the sort param.

* blank