[5.3] Update Icon of Misc Information in Contact (#46067)
Until version 5.2 the icon for the Miscellaneous information is a black circle with the letter "i" in the middle (ie. icon-info-circle).

In version 5.3, the icon has changed to Home icon (ie. icon-home) which is same as that of the Website.

So, this change would differentiate the field identifications.

Pull Request for Issue # 46057.